--- a/package.json
+++ b/package.json
@@ -22,19 +22,13 @@
     "@fortawesome/react-fontawesome": "^0.1.14",
     "bootstrap": "^4.5.3",
     "d3": "3",
-    "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
     "express": "^4.17.1",
     "formik": "^2.2.6",
     "heroku-ssl-redirect": "^0.1.1",
-    "husky": "^6.0.0",
     "jquery": "^3.5.1",
-    "lint-staged": "^11.0.0",
-    "lodash": "^4.17.20",
     "next": "^10.0.7",
     "ordinal": "^1.0.3",
     "popper.js": "^1.16.1",
-    "prettier": "^2.3.0",
     "react": "16.14.0",
     "react-dom": "16.14.0",
     "react-ga": "^3.3.0",
@@ -42,9 +36,15 @@
     "universal-cookie": "^4.0.4"
   },
   "devDependencies": {
+    "@next/bundle-analyzer": "^10.2.3",
     "eslint": "^7.17.0",
     "eslint-config-google": "^0.14.0",
     "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-react": "^7.22.0"
+    "eslint-plugin-prettier": "^3.4.0",
+    "eslint-plugin-react": "^7.22.0",
+    "eslint-plugin-react-hooks": "^4.2.0",
+    "husky": "^6.0.0",
+    "lint-staged": "^11.0.0",
+    "prettier": "^2.3.0"
   }
 }