Do some performance improvements suggested by Lighthouse:
- Improve [Largest Contentful Paint](https://web.dev/lcp/) from 44.7 s to 25.7 s (57%)
- Improve [Time to Interactive](https://web.dev/tti/) from 47.4 s to 27.7 s (58%)

Despite cutting those metrics in half, this actually didn't improve our Lighthouse performance score at all 🙃

Further performance improvements can be made by replacing `d3` with `d3-geo`, `d3-scale`, and `d3-selection`; and replacing `bootstrap` and `jquery` with `react-bootstrap` and `bootstrap.min.css`, but I've sunk a couple of hours into working on that, and honestly, it's probably not worth it.